docker/app/requirements/base.txt
Django==1.8.6
django-mptt
django-taggit
markdown2
beautifulsoup4
Pygments
redis>=2.10.3
django-cors-headers
djangorestframework
django-rest-swagger
django-rest-auth
django-allauth
schedule
pandas
celery
supervisor
flower


# For modiifed python-instagram (in instagram service app)
httplib2
simplejson
six
